Allow storage backend to be configurable
Add configurables so that one need not have Ceph be the storage
backend for Cinder, Cinder-backup, Glance and Nova.

Let Ceph be the default storage backend for Cinder, Cinder backup,
Glance, and Nova. Allow the default setting to be overriden when the
global config, ENABLE_CEPH_{CINDER,C_BAK,GLANCE,NOVA} is set to
False in the local.conf/localrc file.

Closes Bug: #1521278

Change-Id: I3afe7c57d5d56f15fe533ad5b1dbe36ccd369e6b