Fix formatting and indentation
Since we're about to be voting and I expect a few more people to be
interested in the code, I think now is the best time to fix the
indenting.  It's a large patch, but there should be no functional
changes.

I removed one unused function and renamed internal-only function to be
prefixed with '_'.

Change-Id: Ibb9b26f0cc8719be4fec121c1b9aa279938aebd7