Use ceph.com packages if not already configured
For the ceph job to be voting, infra requires that we pull packages from
a pre-configured repository that's mirrored and reliable for gate jobs.
They have arranged for ceph.com packages to be mirrored and will store a
link to those packages in /etc/apt/sources.list.d/ceph.list.  If this
file is present, we should use it.  If not, then we're free to update
the sources list to fetch the latest packages.

This approach is required to regain voting status in the gate.

Change-Id: I5f811c917e6964c33440ed644597c9e485963626