Install Jewel packages of Ceph for manila
The changes in Ceph that were needed to make manila's cephfs_native
driver work are available in the Jewel branch. It's no longer
necessary to use the 'wip-manila' branch.

Also, with Jewel branch, manila's libcephfs client needs to be root
to perform certain manila operations. So set manila's client as root.

Change-Id: I715f79a6801b12522c047b1df4d950c99a1ffad5