Add release key for installing packages from ceph.com
In commit bb773e7285ee909860e98207a38b87ae22cad03a the key for
verifying release packages from ceph.com was removed when support for
the ceph package mirrors was added. However, it is still needed for
non-gate case of fetching packages from ceph.com, so this adds it
back in that case.

Closes-Bug: #1578494

Change-Id: I18f8927f42710e9365c246c55f26f53403965ee1