Fix typo in ruleset parsing
The rule_id line is formatted as:

    "rule_id": 0,

Which means the id integer is the second column of the parsed output,
not the third.

Since a third column doesn't exist, the call to cut returns nothing,
leaving $RULE_ID invalid, which then causes the subsequent ruleset
commands to fail.

We don't see this normally because $CEPH_REPLICAS is set to 1 by
default.

Change-Id: I0cc57f5abc023cb8b2faeea940f28ab54cd290ff
Related-Bug: #1581159