Fix restack problem with libvirt secret
Restacking with ceph plugin fails when trying to install
the cinder user's ceph authorization key in libvirt's list
of secrets because there is already a secret in that list
with the same uuid.  The function _undefine_virsh_secret
is run on cleanup and should have removed that key from
the libvirt secret list.

Fix a typo in the variable name for the uuid that identifies
the authorization key in the secret list so that the
cleanup routine removes the key as intended.

Closes-Bug: #1570665

Change-Id: I19d2c20979dca49d73c427c5b1413229ff39f394