Fix settings for CephFS data and metadata pools
Although the devstack ceph plugin uses variables - CEPHFS_DATA_POOL and
CEPHFS_METADATA_POOL - as arguments to the 'osd pool create' commands
that provision the pools required for manila to use the CephFS Native
backend, the overrides to the default values for these variables sets
both of them to $CEPHFS_CEPH_POOL. This setting is counter-intuitive,
and worse, forces use of a single pool for both data and metadata in the
case both default values are overridden.

Use CEPHFS_DATA_POOL and CEPHFS_METADATA_POOL as the override
variables for the data pool and metadata pool, respectively.

Change-Id: Iae2859388ce5e74b6f532cf2cf7809aaa09746cb
Closes-Bug: 1593829