Create backing disk using $VOLUME_BACKING_FILE_SIZE
The backing disk currently created is 8GB. devstack-gate
sets that to 24GB when running Tempest. We're seeing ceph
job failures due to NoValidHost in the cinder scheduler
because 8GB isn't enough capacity for Tempest runs. So this
change uses the same backing disk size for the ceph jobs as
we get in the default devstack setup, which uses LVM.

Depends-On: I71be308c8373e9ac429b901c374100c6b3c1e59d

Change-Id: I788eefa6c1d427bf51d2d3d40be4abe0336443e7
Closes-Bug: #1645530