Force pip install update of PasteDeploy
The gate-tempest-dsvm-full-devstack-plugin-ceph-ubuntu-xenial job is
currently failing at a rate of 100% since we updated to use the Pike
UCA. The Ceph packages in the Pike UCA are pulling in python-paste and
python-pastedeploy. When devstack pip installs a newer version of
python-paste to satisfy upper-constraints, we end up with Paste
installed from pypi and python-pastedeploy installed from the Ubuntu
archive and Keystone fails to start with the error:

  "ImportError: cannot import name deploy."

This adds a hack to force an update of PasteDeploy from pypi during the
devstack "stack install" phase to work around the problem and get the
job passing again.

Change-Id: Idf28de8faa56c2d387291c7d14ac6d4bf1c74c6a