Add ceph jobs as zuulv3 native in tree
This migrates the legacy jobs to zuulv3 native ones.

While migrating the jobs I got the impressions that they have not
bin maintained recently, which is possible since they are not
voting / experimental.

The regex in the py27 job includes cinder.* tests, and the job
used the all-plugin tox env (which has site-packages enabled) so
my guess is that the job used to run cinder in tree tests.
Those tests are now available in the cinder-tempest-plugin repo
instead, which could be added to these jobs if needed.

Following a similar setup in the cinder-tempest-plugin[0] we add
a file `tempest_blacklist.txt` that can by amended to list tests
that are known to fail for a ceph based deployment.

[0] https://review.openstack.org/#/c/531206/

Co-Authored-By: Jens Harbott <j.harbott@x-ion.de>
Change-Id: I27f85991ca947809d46ae2dee57704e779e6dfb6