Fix the variable for identity version, change default
The names were used for the same entity, so the default was never
applied. Use the variable defined first (and whose name looks
more consistent with the variable normally used to defined
the identity name).

Also, change the default version to 3. As no one seems to be
using it right now on the gates, it should not break anything.

Change-Id: I4cbb104198e2cdaa8e3e8c4b1f1c20dc3e0048b5