Fix installation of PasteDeploy from upstream pkgs (again)
Similar to what happened some time ago for the stable/queens branch, we
are now seeing issues again on master after there was a new Paste
release being made a couple of weeks ago. So we need to perform
the same workaround of forcing pip to install PasteDeploy over
the distro version on the master branch.

Change-Id: I3e98a08e78182bdf5ec09b30e31b98435ec29456
Closes-Bug: 1803143