Disable [validation]/run_validation in tempest.conf
Devstack and tempest default to enable ssh validation in
a lot of API tests which can fail intermittently:

  http://status.openstack.org/elastic-recheck/#1808010

Also see bug 1802971.

Since a ceph backend for storage shouldn't really have
any impact on whether or not ssh is going to work in a
tempest run, this change disables it in the
devstack-plugin-ceph-tempest jobs which should make the
job faster and more stable.

Change-Id: Iea948bcc43315286e5c130485728152d4710bfcb