tempest: re-enable shelve tests for train
The dependent nova fix resolves the issue with rbd and
shelve/unshelve testing in tempest. Re-enabling the test
needs to be branch-specific since devstack-plugin-ceph
is branch-less but the fixes in nova will be branch
specific so we don't want to enable this on any nova
branch that does not have the fix since it would break
the job on those branches.

Related-Bug: #1653953

Depends-On: https://review.opendev.org/457886/

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Change-Id: Ifbfc15033020c430a93574f2726b2279c9be130d