Add stable versions of devstack-plugin-ceph-tempest-py3
devstack-plugin-ceph is branchless and its jobs run on
stable branches. To tests the devstack-plugin-ceph changes
for stable branches, let' add the stable version of
devstack-plugin-ceph-tempest-py3 job in experimental pipeline.

Adding in experimental pipeline as we do not really have to run
those all the time on check gate but can be run on demand.

Change-Id: I9010f0d6226304537adcbe4e3f97250c754a7de9