zuul: Introduce a multinode ceph job
This change introduces a multinode ceph job based on
tempest-multinode-full-py3 to be used to provide live migration and
other multinode coverage initially in the openstack/nova project.

As this job uses ceph for both Nova ephemeral
([libvirt]/imagebackend=rbd) and Cinder volume storage the live block
migration tests are skipped.

Change-Id: I9ffdff44a3ad42ebdf26ab72e24dfe3b12b1ef8b