zuul: Switch devstack-plugin-ceph-tempest-py3 to voting
While the job itself is not currently voting within openstack/nova the
use of the plugin essentially is within the nova-live-migration job.

As such we should really enable voting of this job within this project
with the eventual goal of also enabling it within openstack/nova and
other projects using it.

Change-Id: Id243b92e800e2d6e8b2228fadbd1c72b194bbfb9