Check that ceph-mon service has started to get its version
Add an extra check to verify that ceph-mon service has started
before trying to retrieve its version.

An already existing function is being used, wait_for_daemon

Since we now guarantee that we are using systemctl, we can use
the is-active feature. If the service is running, the exit code
will be 0.

Also, change the command used to get the mon version. We can use
the ceph version command provided by the new cli instead of using
ceph daemon command which requires more privileges.

Change-Id: I2d4f23721107c4b5cb1c09719c55424eecd0c880
Closes-Bug: #1872855