[manila] Fix pre-test-hook to not setup ipv6 always
IPv6 export policy support was added to the Manila
CephFS-via-NFS driver in the Train release. This
devstack plugin is branchless, so we'll need to
check for the relevant branch to setup the ipv6
test harness.

The pre-test and post-test hooks will be deleted soon
when devstack-gate is dropped by manila test jobs
in the Victoria cycle. So this code is temporary
but required to make ceph-via-nfs tests pass
in stable/queens-stein branches.

Change-Id: I91a54a8f72d1d514afbfa769c7e8f2ac21d502dc
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>