[manila] Don't set tempest opt "capability_snapshot_support"
Manila doesn't provide snapshots unless you
assert that you want them via share type
extra-specs. Tests that require "snapshot_support"
do this by creating a share type with that
extra-spec; so we don't need this config
option to be set.

It's actually detrimental when wanting to enable
snapshots on the Ceph driver [1]

[1] https://review.opendev.org/736827/

Change-Id: I41d6aa885e22d3f21150d49ae7e4a8f369cc96f5
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>