Use apt-get to update repos on Ubuntu
The use of apt to refresh repos is discouraged and prints the following
warning when used:

WARNING: apt does not have a stable CLI interface. Use with caution in scripts

The use of apt also appears to break on our Xenial based branches where
we enable the Xenial Queens Ubuntu Cloud Archive repos that are
unsigned. When using apt to refresh these repos apt returns a failure,
while apt-get only returns a warning.

This change simply switches to apt-get to avoid the above warning and to
allow the older Xenial based stable branches to continue to work.

Change-Id: Ic534a47fff17d36eb1ef128ad6af70a34ed942c9