[manila][ci] Disable tempest opt for snapshot cloning
CephFS drivers don't support cloning snapshots
into shares. Manila's tempest tests need to be
told this explicitly. So, update the gate hooks
that control tempest configuration.

This affects all legacy style CephFS jobs that
use devstack-gate. These jobs will go away soon,
and be replaced with Zuulv3 native style job
definitions and this patch is a stop-gap,
because devstack-plugin-ceph is a branchless
devstack plugin, and we need to support legacy
style jobs for a while.

Change-Id: I885ff78586540d01590f98cca2a00d55161dbe04
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>