shaman: use 'ready' repo instead of any
We seem to have been requesting the latest build from the Shaman API
provided by Ceph however we did not filter the status to be 'ready'
which meant that we sometimes got 'building' status repositories that
cause the jobs to fail.

This should stabilize jobs a lot more by using only the ready
repositories.

Change-Id: Iff47bbe4954da3ab582a007ba04b3df50ae21cd4