Stable jobs: run the tests on victoria, remove stein
Update the list of tested stable branches (excluding stein which is EM).
The nodesets for the stable branches are now pinned because
their testing environment are well defined and won't change.

Also, remove the -py3 suffix: those jobs are all -py3 based
(the suffix was mostly requires in the transitional time
between py2 and py3).

Change-Id: I48f8574960df51fc47f80fa17091fd13ec71aa99