Fix job definition as devstack-plugin-ceph is branched
As devstack-plugin-ceph is branched, we moved stable branch
jobs definition in their respective branch

- https://review.opendev.org/c/openstack/devstack-plugin-ceph/+/786308
- https://review.opendev.org/c/openstack/devstack-plugin-ceph/+/786309
- https://review.opendev.org/c/openstack/devstack-plugin-ceph/+/786310
- https://review.opendev.org/c/openstack/devstack-plugin-ceph/+/786311

But we did not cleanup the master version of job definition so that we
do not need to do it for future stable branches. When stable/xena is cut
from master the ./.zuul.yaml version of jobs in stable/xena has the stable
branches variant which is being picked up by zuul for stable branch gate
instead of their own version defined in their stable branch.

- https://zuul.opendev.org/t/openstack/build/bc088154e50440fe9a2cacf52426748a/log/zuul-info/inventory.yaml#241

And it started failing ceph job when ceph version was update for master and
same is being picked by stable branch testin so due to above branch matching
logic.

To avoid this in future when we will have stable/yoga or so we need to cleanup
this in master. Master repo will only keep the master version of jobs and no
stable branch jobs.

Other stable/rocky and older jobs will be moved to stable/stein so that we have
them there if anyone cut the branch.

This also remove the master-only(supposed to run on master gate only)
job 'devstack-plugin-ceph-master-tempest'.

Closes-Bug: #1945358
Change-Id: I7061f8d1491ff957452c9c777e40186a4e9c324e
(cherry picked from commit bf07abf7a9af586b280914fe0f8c00b130ac1b65)