remove python3-logutils
this change ensure that python3-logutils is removed so that
that it can be installed by pip later.

Before today in passing build, there were no python3-logutils pulled by neutron deps so
it was not installed or tried to uninstall so all good
- https://zuul.openstack.org/build/590c5996ca1b402486bfe1c7e1d08535/log/job-output.txt

But from today (10th Dec), python3-logutils pulled by neutron deps and failure started
- https://zuul.opendev.org/t/openstack/build/722c6caf8e454849b897a43bcf617dd2/log/job-output.txt#9419

The root cause of why this issue started happening today is not known. May be its
pecan===1.4.1 ? I8ee467bbb363f428a005f92554812bfdae95881a making it install but there
is no change for logutils as deps in pecan previous version 1.3.3 also[1]).  Or it may
be cpeh/ubuntu packaging.

But it is clear that python3-logutils  is coming from somewhere and causing gate blocker
in Nova gate, let's remove it and later we can find the root cause if anyone need this
package for their ceph job.

[1] https://github.com/pecan/pecan/blob/1.3.3/requirements.txt#L5

Closes-bug: #1954427

Change-Id: Icb63649b252fd6eb229adeae454b5ec3c6b79cad