[stable-only] Pin cinder-tempest-plugin version for stable/train
Currently the devstack-plugin-ceph-tempest(-py3) jobs are failing with
the following errors:

  * AttributeError: 'SnapshotDataIntegrityTests' object has no attribute
    'create_security_group'
  * AttributeError: 'TestEncryptedCinderVolumes' object has no
    attribute 'create_security_group'

These tests reside in the cinder-tempest-plugin, which is branchless.
On stable/train, tempest is pinned to version 26.1.0 [1] and in that
version there is not a create_security_group method in
tempest/scenario/manager.py.

This pins cinder-tempest-plugin to version 1.3.0 which is compatible
with tempest 26.1.0.

[1] https://review.opendev.org/c/openstack/devstack/+/822380

Change-Id: Id11527b17458a77fd0b871279fe7fd099a186242