Temporary pin the ceph jobs nodeset to Focal
As per the community wide goal to migrate the CI/CD from
Ubuntu Focal to Ubuntu jammy, we need to merge the devstack, tox base
jobs to jammy on Nov 18. But ceph jobs are failing on Ubuntu Jammy.

To move ahead to merge the base job patches we need to pin the
ceph jobs nodeset to Focal until this is fixed for Jammy.

Since the devstack-plugin-ceph-tempest-py3 is currently permanently
failing, although this patch is a no-op, make it non-voting. Also don't
run 100% failing non-voting jobs until they are fixed.

Needed-By: https://review.opendev.org/c/openstack/devstack/+/860795

Related-bug: #1996628
Change-Id: Id2ae61979505de5efb47ce90a2bac8aac2fc5484