Cap cinder-tempest-plugin version for stable/wallaby
As stable/wallaby is in EM phase, and currently broken
with the Tempest master. We are pinning tempest there (depends-on).

To have compatible cinder-tempest-plugin version for stable/wallaby
ceph jobs, we need to pin cinder-tempest-plugin 1.7.0 version.

Also, we need the os-brick fix
- https://review.opendev.org/c/openstack/os-brick/+/864479
As wallaby is in EM and os-brick cannot be released with the above fix,
we need to install it from stable/wallaby git version instead of pip
from u-c.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/871782

Change-Id: Ie47c3e9bb41e3513f65735a657e7a3a5dbfa27e5