Add service secret to libvirt when cinder enabled
The current code only adds the service secret to libvirt when Ceph has
been enabled in Nova, but it should also be enabled if it has only been
enabled in Cinder.

This patch changes devstack/plugin.sh to write the service secret to
libvirt whenever Nova or Cinder is using Ceph.

This would be the counterpart of the code we already have in
_undefine_virsh_secret where we are already checking if it's enabled in
either of the services to remove the secret.

Change-Id: I1067d52b7a435fcef7996eea6479d598be842dca