[Partial Backport] Revert "Temporary pin the ceph jobs nodeset to Focal"
Note(sean-k-mooney):
This is a partail backport of I899822fec863f43cd6c58b25cf4688c6a3ac1e9b
contianing only the change to enable validations in the base job
and the swap/concurrency/mysql changes to account for the high memory
pressure in the job which leads to instablity. All changes outside
of the .zuul.yaml change are dropped as is the depend on for the
cinder-tempest-plugin. cinder-tempest-plugin is branchless
so we do not need to backport it and it is already merged on master
so the depency is fulfilled.

This reverts commit 863a01b03286e6595d68ac7f2560c857bcf944c5.

Partial revert only for the pin to focal, leaves the broken other jobs
commented out.

Update paste-deploy workaround to be used always.
Add qemu-block-extra and podman deps to the debs list.
Running on the newer ceph and distro causes some quite different
performance characteristics that cause tests that used to pass to fail
more often. This includes some performance optimizations to help
reduce the memory footprint, as well as depends on changes to
tempest tests to improve the reliability of those tests by enabling
validation via SSH.

This also moves the cephadm job to be the voting/gating job as that
seems to be the clear consensus about "the future" of how we deploy
ceph for testing.

Co-Authored-By: Dan Smith <dms@danplanet.com>
Change-Id: I899822fec863f43cd6c58b25cf4688c6a3ac1e9b
(cherry picked from commit 41b6a8c227190a9b52d29a078425321f96240d92)
(cherry picked from commit 2d61c38586bfd25cc6cf858c52c739f44f96c191)