Set CephFS/NFS job to non-voting
This job is non-voting against the manila repos currently.

We've observed sporadic failures in data intensive scenario
tests. There are a number of manila tests that write a few
gigabytes of data onto mounted shares. While we've tried
to optimize test concurrency and the kind of data being written,
sometimes, things seem to fall over and a test case runs into
a generic I/O error. So maybe more work is necessary to
get this job to be stable; but that work may not be in this
devstack plugin repository.

Change-Id: Id89f161ba6dd5b5fbb5b82042a3e94c92797a881
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>