[manila] Enforce cephfsnative as the default driver
During the devstack setup, if you have set `CEPHADM_DEPLOY=True`,
we are not considering cephfsnative as the default driver, even
though it was the expected behavior before.

If we do not set this flag and it is not set in the local.conf,
the devstack setup fails.

That is fixed by enforcing `cephfsnative` as the default value for
the Ceph backend. In case users want to deploy devstack with Ceph
NFS, they can continue using the same workflow (setting
`MANILA_CEPH_DRIVER=cephfsnfs`).

Change-Id: I51660fa2466fff873f2230e683661b53874bf862
Signed-off-by: Carlos Eduardo <ces.eduardo98@gmail.com>