Fix lacking ceph ingress IP on devstack
When the Ceph ingress IP is not set, devstack is currently failing
to install with CephFS NFS. When it attempts to create the NFS
cluster, it fails because the VIP value is not set.

The VIP should be the same as the ceph ingress IP, but it is only
being set in the manila job definitions, so if you don't provide
it in the local.conf file manually, the value will be empty.

This change fixes it by defaulting the ceph ingress IP to the
host IP in case it wasn't provided in the local.conf.

Change-Id: Ib2db0faa5381da9e3d391ff0f887eb92dff9c295
Signed-off-by: Carlos Eduardo <ces.eduardo98@gmail.com>