Set osd memory config
Still seeing oom-killer trigger in Cinder Ceph CI
jobs with ceph-osd being a top memory consumer.

It appears to default to trying to use 4GB of memory,
change this to 2GB to see if it helps.

Increases the job timeout to 3h for a bit more
headroom since jobs are regularly taking more than 2h
to complete.

Change-Id: I71a46a452914256f36299623a29ed7eebf8e61d2