Deploy rbd-mirror daemon
It is possible to setup one way or two way replication between pools.
rbd-mirror daemon is responsible for pulling image updates from the
remote peer cluster and applying them to the image within the local
cluster. This patch allows to enable this daemon as part of the
deployment, but the actual configuration between sites remains out
of scope for the deployment script.
This work aims to support [1].

[1] https://review.opendev.org/c/openstack/cinder-tempest-plugin/+/948293

Change-Id: Id1a66c244b9be33b6df63e6504500c77d59a1b9d
Signed-off-by: Francesco Pantano <fpantano@redhat.com>